requirements.txt
duckdb>=0.9.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.1
google-cloud-bigquery>=3.11.0
pyarrow>=12.0.0
python-dotenv>=1.0.0
